Modified trc_sbc_bfm for NEMO
This version restores the usage of river input.
Dilution from precipitation and concentration from
evaporation are computed for all variables.
River dilution is only computed for variables that have some input
from river, controlled by the flag in the bfm.nml namelist and
specified in the file namelist_top. For all the other variables
it is assumed that the river carries exactly the same concentration
as found in the first ocean grid point (this removes the high
dilution found in high resolution simulations).
If the runoff and river load masks do not correspond, this may create
a local imbalance, because there can be dilution in one grid point
and input in another one (it is the case in the standard PELAGOS).

Also corrected a log message in ModuleCO2 that put information
on std output.