Improved definition of analytical light (#103)
There used to be a mismatch between analytical light forcing and
computation of primary producyion in Phyto.F90.
The parameter LightForcingFlag was set in Param.nml but also
controlled the way analytical light was computed (see #103)
Now two new parameters are defined, one for setting the analytical
forcing function ltype and one for the production, renamed
ProductionLightFlag.