Improved build/config_STANDALONE.sh
The configuration of STANDALONE now generates the BFM files in
the BLD_STANDALONE directory (as done with NEMO).
This simplify the diagnosis of problems from users when they are due
to tclsh generation.
The namelists of the standard tests have also been updated to the last
version of the code as they were out of sync.