Remove specific BentoPelCoup.F90 for STANDALONE
It is now possible to use the same BentoPelCoup.F90 for the
STANDALONE and for the other couplings.
The STANDALONE is now by default a closed box because
AssignPelBenFluxinBFM is set to FALSE in the default namelist.
If set to TRUE the interface with the benthic system is turned on
and mass is removed from the bottom