Make the calls to plankton dynamics uniform
Phytoplankton dynamics was called with one single argument with
the phyto group, while all the other functional groups were
called with the arguments for C, N and P components.
Now they are all equal and the access to the components
is done inside the specific routine.
The description in the manual is in agreement with this method.