A new data output handling was created to allow monthly output and solve bug #285.
A new subroutine was generated, namely update_delta_save (in api.bfm.F90),
which nternally set on the basis of the out_delta parameter of BFM_General.nml
the output stepping to save the data. If a negative value is given to out_delta
then real monthly data are stored. The routine pad the output step to end of the
simulation (bug #285).
NOTE: the criteria to save the output is not based anymore on the MOD function,
but it compares the values of the actual timestep and the output timestep.