Update seaice bacteria and fauna
The dynamics of sea ice bacteria and fauna are now compliant with
the pelagic counterparts. Bacteria have only one parameterization
(BACT1, Vichi et al. 2007) as there is no semi-refractory DOC
in sea ice yet. Fauna parameterizations are taken from microzooplankton.

Note that at the moment SeaicetoPelCoup.F90 does not work correctly for
the silica component of detritus as there is a bug in the generation of
the code (#389).