Move trcbc to the NEMO directory
This work is done in preparation to NEMO3.6. The versions
of trcdta and trcbc that are found in this directory can be
restored in case the used version of NEMO does not include them.
We may decide to put them here when it is clear that they are
compatible with all versions from NEMO3.4 to the future 3.6