Commit related to the Feature #438 (Bnmerge in parallel)
Parallel execution is not possible currently. The problem is the libraries which BNMERGE use, they are not compatible with fortran. Specifically, is not possible to compile HDF5 with options "enable-threadsafe"  and "enable-fortran", so it is not possible to run OMP safely and then parallel execution fails because HDF5 errors.

If in the future HDF5 accepts fortran and threadsafe, it is possible to test the program compiling with the environment variable "PARALLEL=yes". In the meantime, parallel compiling and execution is hidden for the users.

Comment to add to README if threadsafe parallel is possible:
-To execute in parallel:
-   - export variables:
-      - PARALLEL equal to yes: export PARALLEL=yes
-      - OMP_NUM_THREADS to number of parallel threads: export OMP_NUM_THREADS=8
-   - load libraries:
-      - hdf5-1.8.11_parallel
-      - netcdf-4.3_parallel (with "netcdf-fortran-4.2")
-      - parallel-netcdf-1.3.1