Solve Bug #435: remove INCLUDE_PELCO2 key gives segmentation fault in FluxFunctions
FluxFunctions was taken the arguments as "input" parameters. When one of these parameters
("origin" or "destination") have value of 0, Fluxfunctions modify the value of the parameter.
If key INCLUDE_PELCO2 is removed. Subroutine PhytoDynamics call FluxFunctions with origin
declared as parameter. Because a parameter cannot be modified, an error occurred.

Solution was to copy the parameters to local variables inside Fluxfunctions, so they
can be modified inside the subroutine without any problem