Solve output wrong when key INCLUDE_PELCO2 is not actived (realted Bug #435)
- flux_vector modify the input "flux" changing the sign, but this behaviour
only produces wrong output when INCLUDE_PELCO2 was not activated. It have
been implemented a sign restorei at the end of flux_vector when modified.

- Also some correctios in if-then logic have been made in the function
"fixed_quota_flux_vector". Removed initialization when origin or destination
equal to 0, because it is made inside "flux_vector" function ( duplicate
initialization )

- Thanks Tomas!! Tomas has been the malefic mind behind these changes, I've
been merely a code writer and a "git" committer