Solve Bug #450 . 2D Surface incorrect values in merging with bnmerge
There is a problem with the function "nf90_get_var". It reads incorrect
values if the array used to store the data is bigger than the size of the
specified data to read from the file (start,count). The array was allocated
with the maximun size between oceanpoints, bottompoint, surfacepoint, this
produces incorrect reading in "nf90_get_var". The solution is to allocate the
array with the exact size of the dimension to read.