Modified restart I/O handling (#480) and improved bfm.log structure (#481)
A new restart I/O handling was developed and a new option unpad_out
was created to save restart and output data simualtaneosuly.
Use of unpad_out prevent the padding of data at the end of simulation.
Started to improve the content of bfm.log (#481).
Cosmetics to some files and some modification to update_delta function.