Make trc_nxt_bfm default with NEMO
The BFM is supposed to be used only with the Euler
integration (ln_top_euler).
There was an old option to use the standard trc_nxt
from NEMO and the one from the BFM that also contains the hacked
open boundary conditions.
This part needs to be revised when BDY is implemented