Correction of the statistics computation for disabled variables in NEMO-BFM coupling.
When Functional types are excluded from the execution of BFM,
the related summary statistics computed after the transport step
showed the same values of the previous active variables.
This misbehaviour may induce to think that the code has a bug.
The code was corrected to shoe the statistics of the right variable.