Revision of transport routine for NEMO3.6 coupling
Note that until the nemo work-flow on passive tracers will change,
the partial step computation (zps_hde) in bfm must be done before
lateral diffusion (trc_ldf) to correctly pass arrays on bottom gradients.
In nemo 3.6, the bottom gradients are computed for all tracers,
stored in gtru and gtrv and passed to ldf in the next time step.