Add automatic generation of nutrients-to-carbon default paramters in init_var_bfm (Bug #509)
Changes to print_f90.pm script in build/scripts/bin to include the automatic generation
of the code to set the default nutrients-to-carbon ratio parameters in init_var_bfm,
when a new constituent is added in the layout (Bug #509).
The new functon in the parser, set the Redfield ratios for nitrate, phosphate and silica,
while chla and iron ratios are set to reference values of diatoms.
All the other additional constituents are set to ZERO.