Add missing constituent ratios assignment generation for other components in init_var_bfm.
bfm_onfigure was not generating for the inclusion of carbon ratios for additional parameters (e.g. iron).
Changes to perl script printf90.pm were made to allow the prescriprition of ratios provided
in namelists_bfm using the name convention GROUP_parameters*, where the * is none or "_CONSTITUENTNAME".
Note that, the PelDetritus group is still initialized with prescribed default values,
but this can be modified by adding in the code and in model namelists (using PelDetritus_parameters)
the specific carbon-to-constituent ratios (e.g., p_qncOMT).